stage changes as js

+++ contracts/Orchestrator.sol
@@ -4,21 +4,19 @@ import "openzeppelin-eth/contracts/ownership/Ownable.sol";
 
 import "./UFragmentsPolicy.sol";
 
-
 /**
  * @title Orchestrator
  * @notice The orchestrator is the main entry point for rebase operations. It coordinates the policy
  * actions with external consumers.
  */
 contract Orchestrator is Ownable {
-
     struct Transaction {
         bool enabled;
         address destination;
         bytes data;
     }
 
-    event TransactionFailed(address indexed destination, uint index, bytes data);
+    event TransactionFailed(address indexed destination, uint256 index, bytes data);
 
     // Stable ordering is not guaranteed.
     Transaction[] public transactions;
@@ -41,18 +39,15 @@ contract Orchestrator is Ownable {
      *         If a transaction in the transaction list reverts, it is swallowed and the remaining
      *         transactions are executed.
      */
-    function rebase()
-        external
-    {
-        require(msg.sender == tx.origin);  // solhint-disable-line avoid-tx-origin
+    function rebase() external {
+        require(msg.sender == tx.origin); // solhint-disable-line avoid-tx-origin
 
         policy.rebase();
 
-        for (uint i = 0; i < transactions.length; i++) {
+        for (uint256 i = 0; i < transactions.length; i++) {
             Transaction storage t = transactions[i];
             if (t.enabled) {
-                bool result =
-                    externalCall(t.destination, t.data);
+                bool result = externalCall(t.destination, t.data);
                 if (!result) {
                     emit TransactionFailed(t.destination, i, t.data);
                     revert("Transaction Failed");
@@ -66,25 +61,15 @@ contract Orchestrator is Ownable {
      * @param destination Address of contract destination
      * @param data Transaction data payload
      */
-    function addTransaction(address destination, bytes data)
-        external
-        onlyOwner
-    {
-        transactions.push(Transaction({
-            enabled: true,
-            destination: destination,
-            data: data
-        }));
+    function addTransaction(address destination, bytes data) external onlyOwner {
+        transactions.push(Transaction({enabled: true, destination: destination, data: data}));
     }
 
     /**
      * @param index Index of transaction to remove.
      *              Transaction ordering may have changed since adding.
      */
-    function removeTransaction(uint index)
-        external
-        onlyOwner
-    {
+    function removeTransaction(uint256 index) external onlyOwner {
         require(index < transactions.length, "index out of bounds");
 
         if (index < transactions.length - 1) {
@@ -98,10 +83,7 @@ contract Orchestrator is Ownable {
      * @param index Index of transaction. Transaction ordering may have changed since adding.
      * @param enabled True for enabled, false for disabled.
      */
-    function setTransactionEnabled(uint index, bool enabled)
-        external
-        onlyOwner
-    {
+    function setTransactionEnabled(uint256 index, bool enabled) external onlyOwner {
         require(index < transactions.length, "index must be in range of stored tx list");
         transactions[index].enabled = enabled;
     }
@@ -109,11 +91,7 @@ contract Orchestrator is Ownable {
     /**
      * @return Number of transactions, both enabled and disabled, in transactions list.
      */
-    function transactionsSize()
-        external
-        view
-        returns (uint256)
-    {
+    function transactionsSize() external view returns (uint256) {
         return transactions.length;
     }
 
@@ -123,12 +101,10 @@ contract Orchestrator is Ownable {
      * @param data The encoded data payload.
      * @return True on success
      */
-    function externalCall(address destination, bytes data)
-        internal
-        returns (bool)
-    {
+    function externalCall(address destination, bytes data) internal returns (bool) {
         bool result;
-        assembly {  // solhint-disable-line no-inline-assembly
+        assembly {
+            // solhint-disable-line no-inline-assembly
             // "Allocate" memory for output
             // (0x40 is where "free memory" pointer is stored by convention)
             let outputAddress := mload(0x40)
@@ -142,14 +118,12 @@ contract Orchestrator is Ownable {
                 // + callValueTransferGas (9000) + callNewAccountGas
                 // (25000, in case the destination address does not exist and needs creating)
                 sub(gas, 34710),
-
-
                 destination,
                 0, // transfer value in wei
                 dataAddress,
-                mload(data),  // Size of the input, in bytes. Stored in position 0 of the array.
+                mload(data), // Size of the input, in bytes. Stored in position 0 of the array.
                 outputAddress,
-                0  // Output is ignored, therefore the output size is zero
+                0 // Output is ignored, therefore the output size is zero
             )
         }
         return result;
